Partial work on refactoring ListPanel component.
The plan is to make the ListPanel more of a functional component where
only the layout, design and interaction are handled by the ListPanel.
Any actual implementation will include it as a component, pass it
data like the title, any actions, filters, etc, and receive events
like when search phrase changes, and execute those events in the
parent component.

The goal is to make it easier to make a simple ListPanel, where you
just define the actions you and data, pass it into the ListPanel,
and you can work with it that way. This should minimize the amount
of custom template work required for ListPanels. BUt it will mean
these ListPanels, like Announcements, will have to define and pass
information about the buttons they want to add and what they want
the buttons to do.